requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
httpx>=0.27,<1.0
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
supabase>=2.3,<3.0
google-genai>=0.4,<1.0

